requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pyodbc>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=0.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
pyarrow>=10.0.0
streamlit>=1.30.0